Rename files to match change in name
